Opensora-hpcai: support modelarts training (#489)
* convert to abs path, add MA training scripts

* small fix on infer_vae.py

* small fix

* modify dtype in infer_pipeline.py

* precommit

* move modelarts scripts to mindone/tools

+++ examples/opensora_hpcai/opensora/pipelines/infer_pipeline.py
@@ -56,7 +56,7 @@ class InferPipeline(ABC):
     def vae_encode(self, x):
         image_latents = self.vae.encode(x)
         image_latents = image_latents * self.scale_factor
-        return image_latents.astype(ms.float16)
+        return image_latents
 
     def vae_decode(self, x):
         """
